Refactor semantic module: split assembler and conversion
- Deleted orphan file `src/cli.rs`.
- Refactored `src/semantic/assembler.rs` into `src/semantic/assembler/` directory with `mod.rs`, `state.rs`, and `checks.rs`.
- Refactored `src/semantic/conversion.rs` into `src/semantic/conversion/` directory with submodules for assertions, bindings, collections, expressions, functions, iterators, and print.
- Updated imports and verified tests pass.
- Ensured YAGNI by removing unused code during refactoring.
