+++ src/semantic/conversion/bindings.rs
@@ -1,0 +1,132 @@
+use super::expressions::extract_value;
+use crate::errors::GlossaError;
+use crate::morphology;
+use crate::semantic::assembler::state::AssembledStatement;
+use crate::semantic::{AnalyzedExpr, AnalyzedExprKind, AnalyzedStatement, Scope};
+use crate::text::normalize_greek;
+
+/// Helper: Detect variable binding (let x = ...)
+pub fn classify_variable_binding(
+    asm_stmt: &AssembledStatement,
+    scope: &mut Scope,
+) -> Result<Option<AnalyzedStatement>, GlossaError> {
+    if let Some(ref verb) = asm_stmt.verb {
+        let verb_lemma = normalize_greek(&verb.lemma);
+
+        if crate::morphology::lexicon::is_binding_verb(&verb_lemma) {
+            let has_false_participle = !asm_stmt.participles.is_empty()
+                && morphology::lexicon::lookup(&asm_stmt.participles[0].verb_lemma).is_none();
+
+            let (var_name, actual_asm) = if has_false_participle {
+                let first_participle = &asm_stmt.participles[0];
+                let mut fixed_asm = asm_stmt.clone();
+                fixed_asm.participles = asm_stmt.participles[1..].to_vec();
+                (normalize_greek(&first_participle.original), fixed_asm)
+            } else if let (Some(subject), Some(object)) = (&asm_stmt.subject, &asm_stmt.object) {
+                let subject_name = normalize_greek(&subject.original);
+                let object_name = normalize_greek(&object.original);
+
+                if scope.is_defined(&subject.lemma) && !scope.is_defined(&object.lemma) {
+                    let mut swapped = asm_stmt.clone();
+                    swapped.subject = Some(object.clone());
+                    swapped.object = Some(subject.clone());
+                    (object_name, swapped)
+                } else {
+                    (subject_name, asm_stmt.clone())
+                }
+            } else if let Some(subject) = &asm_stmt.subject {
+                (normalize_greek(&subject.original), asm_stmt.clone())
+            } else if !asm_stmt.participles.is_empty() {
+                let first_participle = &asm_stmt.participles[0];
+                let mut fixed_asm = asm_stmt.clone();
+                fixed_asm.participles = asm_stmt.participles[1..].to_vec();
+                (normalize_greek(&first_participle.original), fixed_asm)
+            } else {
+                return Err(GlossaError::semantic("Binding without subject"));
+            };
+
+            let (value_expr, value_type) = extract_value(&actual_asm, scope)?;
+
+            let final_value_expr = if asm_stmt.is_propagate {
+                AnalyzedExpr {
+                    glossa_type: value_type.clone(),
+                    expr: AnalyzedExprKind::Try(Box::new(value_expr)),
+                }
+            } else {
+                value_expr
+            };
+
+            let is_mutable = asm_stmt.has_mutable_marker;
+            if is_mutable {
+                scope.define_mut(var_name.clone(), value_type.clone());
+            } else {
+                scope.define(var_name.clone(), value_type.clone());
+            }
+
+            return Ok(Some(AnalyzedStatement::Binding {
+                name: var_name.clone(),
+                value: final_value_expr,
+                mutable: is_mutable,
+            }));
+        }
+    }
+    Ok(None)
+}
+
+/// Helper: Detect variable assignment (x = ...)
+pub fn classify_assignment(
+    asm_stmt: &AssembledStatement,
+    scope: &mut Scope,
+) -> Result<Option<AnalyzedStatement>, GlossaError> {
+    if let Some(ref verb) = asm_stmt.verb {
+        let verb_lemma = normalize_greek(&verb.lemma);
+
+        if crate::morphology::lexicon::is_assignment_verb(&verb_lemma) {
+            let var_name = if let Some(ref subject) = asm_stmt.subject {
+                normalize_greek(&subject.original)
+            } else {
+                return Err(GlossaError::semantic("Assignment without subject"));
+            };
+
+            let binding = scope.lookup_binding(&var_name);
+            match binding {
+                None => {
+                    return Err(GlossaError::semantic(format!(
+                        "Τὸ «{}» οὐχ ὡρίσθη — πρῶτον ὅρισον αὐτό",
+                        var_name
+                    )));
+                }
+                Some(b) if !b.mutable => {
+                    let _b = b; // Silence unused variable warning while keeping guard
+                    return Err(GlossaError::semantic(crate::errors::immutable_assignment(
+                        &var_name,
+                    )));
+                }
+                Some(_b) => {
+                    let has_value = !asm_stmt.literals.is_empty()
+                        || asm_stmt.object.is_some()
+                        || !asm_stmt.arrays.is_empty()
+                        || !asm_stmt.unwraps.is_empty()
+                        || !asm_stmt.index_accesses.is_empty()
+                        || !asm_stmt.property_accesses.is_empty()
+                        || !asm_stmt.nested_phrases.is_empty();
+
+                    if !has_value {
+                        return Err(GlossaError::semantic(format!(
+                            "Τῇ πράξει «{} γίγνεται» δεῖ τιμῆς (Assignment requires a value)",
+                            var_name
+                        )));
+                    }
+
+                    let (value_expr, _) = extract_value(asm_stmt, scope)?;
+                    scope.mark_used(&var_name);
+                    return Ok(Some(AnalyzedStatement::Assignment {
+                        name: var_name.clone(),
+                        value: value_expr,
+                    }));
+                }
+            }
+        }
+    }
+    Ok(None)
+}
